Update des fichiers et meilleure organisation
